Cleanup code
- clean declaration and initialization of constants
- avoid implicit type casts
- add 'this' operator where it was missing

+++ libCompanion/companion/algo/2D/ShapeDetection.h
@@ -53,7 +53,7 @@ namespace Companion
 			ShapeDetection(cv::Mat morphKernel = cv::getStructuringElement(cv::MORPH_RECT, cv::Size(30, 30)),
 				cv::Mat erodeKernel = cv::getStructuringElement(cv::MORPH_RECT, cv::Size(10, 10)),
 				cv::Mat dilateKernel = cv::getStructuringElement(cv::MORPH_RECT, cv::Size(40, 40)),
-				int cannyThreshold = 50,
+				double cannyThreshold = 50.0,
 				int dilateIteration = 3);
 
 			virtual ~ShapeDetection();
@@ -77,7 +77,7 @@ namespace Companion
 			/**
 			 * Canny threshold value to obtain edges.
 			 */
-			int cannyThreshold;
+			double cannyThreshold;
 
 			/**
 			 * Morphology transformation kernel for morphologyEx operation.
